fix Link issues and consolidate middleware
