+++ components/layout.tsx
@@ -15,10 +15,8 @@ type AsteriskProps = {
 const AsteriskLink: React.FC<AsteriskProps> = ({ spinning }) => {
   return (
     <>
-      <Link href="/">
-        <a title="Home">
-          <Asterisk />
-        </a>
+      <Link href="/" title="Home">
+        <Asterisk />
       </Link>
       <style jsx>{`
         a {
@@ -122,7 +120,7 @@ const Layout: React.FC<Props> = ({
                   <div className="footer-link info"><a role="presentation" onClick={() => setModalOpen(true)}>Info</a></div>
                   <div className="footer-link mux">An <a href={OPEN_SOURCE_URL}>open source</a> project by <a href={MUX_HOME_PAGE_URL}>Mux</a></div>
                   <div className="divider" />
-                  <div className="footer-link terms"><Link href="/terms"><a>Terms</a></Link></div>
+                  <div className="footer-link terms"><Link href="/terms">Terms</Link></div>
                   </>
   
                 }
